Added Dashboard tables
