+++ package.json
@@ -13,6 +13,7 @@
     "react": "^17.0.2",
     "react-bootstrap": "^2.1.2",
     "react-chat-widget": "^3.1.4",
+    "react-datepicker": "^4.8.0",
     "react-dom": "^17.0.2",
     "react-redux": "^8.0.1",
     "react-router-dom": "^6.2.2",
