+++ src/components/Common.css
@@ -71,7 +71,7 @@
 
 /* header */
 .HeaderBlock{
-    background: url('https://media.istockphoto.com/photos/ballpoint-pen-resting-on-top-of-stack-of-documents-picture-id1012160724?k=20&m=1012160724&s=612x612&w=0&h=k9oh7gKaJl9PPGJMytz9crhewrTB0A-BWb3bHoWRIIg=') no-repeat;
+    background: url('background.jpg') no-repeat;
   background-position: 50% 100%;
   background-size: cover;
   background-position-y: center ;
@@ -101,7 +101,24 @@
     padding-top: 5%;
     opacity: 100%;
     color: #000000;
+    background-position-x: left;
   }
 .container{
     align-content: center;
+    background-position-x: left;
+}
+
+/* Cards */
+.site-card-wrapper{
+    margin: 0 0 30px;
+    color: #109CF1;
+    padding-left: 10%;
+    padding-right: 10%;
+}
+
+.WidgetLg{
+    flex:2;
+    -webkit-box-shadow: 1px -4px 11px -4px #000000; 
+    box-shadow: 1px 1px 2px 1px #c0bdbd;
+    padding: 20px;
 }
