+++ src/components/Dashboard/Header.js
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from 'react'
-import { Row, Col } from 'antd';
+
 
 function Header() {
     var [date, setDate] = useState(new Date());
@@ -12,12 +12,6 @@ function Header() {
 
     return (
         <div className="HeaderBlock">
-            <div className="container">
-                <Row>
-                    <Col span={16}><h1>Welcome to DocTracker</h1></Col>
-                    <Col span={8}><h4><p> {date.toLocaleTimeString()}</p><p>  {date.toLocaleDateString()}</p> </h4></Col>
-                </Row>
-            </div>
         </div>
     );
 }
