+++ src/screens/Dashboard.css
@@ -1,0 +1,14 @@
+.small{
+    flex: 1;
+    -webkit-box-shadow: 1px -4px 11px -4px #000000; 
+    box-shadow: 1px 1px 2px 1px #c0bdbd;
+    padding: 20px;
+    margin-right: 20px;
+}
+
+.large{
+    flex:2;
+    -webkit-box-shadow: 1px -4px 11px -4px #000000; 
+    box-shadow: 1px 1px 2px 1px #c0bdbd;
+    padding: 20px;
+}
